Add simple report flag to dweets and comments
This adds simple report flags to dweets and comments. Reports are sent
to moderators on Discord. Injecting more complicated HTML through
js/ajax for the "load more comments" section turned out to be a little
complex w.r.t. CRSF. Then I realized that we rarely have enough comments
to trigger a "Load more comments" link, and when it is there, it often
hides just a single comment. Therefore, I figured we can make the "Load
more comments" link be a simple view-level thing and not actually do any
ajax loading.

In order to smoothly place the report button, I did a small visual
refresh of the comments as well.